Use the build location specified in preferences (which is just the default).
Added post-action scripts to copy the executables into /Local.  The "all" target needs one too, since
it doesn't execute post-action scripts for the dependant targets.  This provides a convenient short
path to use when I'm testing in Terminal.

Prior to Xcode 4.2 or 4.1, I was able to just set this up from the Xcode preferences, and this didn't
affect anyone else.  When that method stopped working, I just changed SYMROOT to what I wanted, but
this is unexpected to other users.  Now it is isolated once again in user-specific settings.