Fix https://github.com/nygard/class-dump/issues/13.
If a property is listed as having the modifier "retain" or "copy",
but is not a garbage-collected type, then we need to decorate it with
__attribute__((NSObject)) to keep the compiler quiet.

Garbage-collected types, as far as I can tell, are only "Class", "id",
block types, and named types such as "InterfaceType *". Despite the
explicit pointer syntax in "InterfaceType *", the Objective-C encoding
of that type is

    @"myInterfaceType"

If it were a pointer to a C struct type instead, it would be encoded as

    ^{myStructType=}

and would not be considered a garbage-collected type.