Don't duplicate struct definitions for Objective-C++ templated types.
The "// Template types" section of the generated CDStructures.h contains
typedef declarations such as

    typedef struct S<int> {
        int _field1;
    } S_9abcdef0;

This causes compiler errors even *after* flattening "S<int>" into the
Objective-C-friendly "S_int_", because the S<int> type was already
defined earlier in the file. What we really want to see here is simply

    typedef struct S<int> S_9abcdef0;

I'm not at all confident that this change is the best way to guarantee
this output, but it does seem to work on my test case.