Correctly pass through the argument to -[NSCopying copyWithZone:].
This is pretty much just a cosmetic change, but the whole point of the
"zone" parameter is that you're supposed to pass it to allocWithZone:,
so let's do that.

-[CDType copyWithZone:] should also use the "zone" parameter, but
in fact it doesn't directly allocate the copied type; it simply
passes along the return value of -[parser parseType:], which is
allocated in an unknown zone.