Add a poorly tested --mangle-template-types option.
This option is designed to do inside class-dump what I've been doing
for a long time outside it with a sed script (that has finally met
its match in a deeply nested template argument).

The output of "class-dump --mangle-template-types" should be identical
to the output without "--mangle-template-types", except that each
type-identifier will have all its non-alphanumeric characters
replaced with underscore (_) characters. This will allow the output
of "class-dump random-objective-c++-binary" to compile as Objective-C
with fewer hacks. (Some hacks are still needed: for example, Apple's
QuartzCore binary has a struct with a field named "_vptr$GenericContext".
I don't know how to reproduce this; it might be a quirk of Apple's
original source code rather than a quirk of the Objective-C++
implementation.)

Right now the output of "--mangle-template-types" is not quite
identical to the old output; it's missing some lines of the form

    typedef struct vector<int> vector_5a588309;

I don't understand why this should be.